pyproject.toml
[tool.poetry]
name = "validated-relationships-service-api"
version = "0.0.1-alpha"
description = "TODO"
authors = [
    #TODO add authors
]
readme = 'README.md'
license = "MIT"
repository = "https://github.com/NHSDigital/validated-relationships-service-api"
homepage = "https://digital.nhs.uk/developer/api-catalogue"
keywords = ["healthcare", "uk", "nhs"] #TODO add additional keywords

[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.group.ci.dependencies]
flake8 = "^3.7.9"
black = "^24.3.0"

[tool.poetry.group.test.dependencies]
pytest-nhsd-apim = "^3.3.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
black = "^24.3.0"
pip-licenses = "^4.4.0"
jinja2 = "^3.1.3"
pyyaml = "^6.0.1"
docopt = "^0.6.2"
jsonpath-rw = "^1.4.0"
semver = "^3.0.2"
gitpython = "^3.0.5"
pytest = "^8.2.0"
coverage = "^7.5"
aiohttp = "^3.10.2"
pytest-asyncio = "^0.14.0"
openapi-schema-validator = "^0.6.2"
